MSVC fix for _ENABLE_EXTENDED_ALIGNED_STORAGE
* Fix a MSVC binary-breaking compatibility in MSVC 2017 15.8:

Fixes errors like:
error C2338: You've instantiated std::aligned_storage<Len, Align> with 
an extended alignment (in other words, Align > alignof(max_align_t)). 
Before VS 2017 15.8, the member type would non-conformingly have an 
alignment of only alignof(max_align_t). VS 2017 15.8 was fixed to handle 
this correctly, but the fix inherently changes layout and breaks binary 
compatibility (*only* for uses of aligned_storage with extended 
alignments). Please define either (1) _ENABLE_EXTENDED_ALIGNED_STORAGE 
to acknowledge that you understand this message and that you actually 
want a type with an extended alignment, or (2) 
_DISABLE_EXTENDED_ALIGNED_STORAGE to silence this message and get the 
old non-conformant behavior.

* uniform_int_distribution fix for msvc
Fixes: 
`error C2338: invalid template argument for uniform_int_distribution: N4659 29.6.1.1 [rand.req.genl]/1e requires one of short, int, long, long long, unsigned short, unsigned int, unsigned long, or unsigned long long `

+++ libs/io/src/CFileGZStreams_unittest.cpp
@@ -21,14 +21,15 @@ namespace mrpt
 extern std::string MRPT_GLOBAL_UNITTEST_SRC_DIR;
 }
 
-const size_t tst_data_len = 1000;
+const size_t tst_data_len = 1000U;
 
 // Generate random data:
 void generate_test_data(std::vector<uint8_t>& tst_data)
 {
-	const int random_seed = 123;
+	const unsigned int random_seed = 123U;
 	std::mt19937 mersenne_engine{random_seed};
-	std::uniform_int_distribution<uint8_t> dist{0, 3};  // low entropy
+	// MSVC C++11 library enforces use of `unsigned short` as smallest type
+	std::uniform_int_distribution<unsigned short> dist{0, 3};  // low entropy
 	auto gen = [&dist, &mersenne_engine]() { return dist(mersenne_engine); };
 	// auto gen = []() { return 1; };
